// frontend/src/client/models/AnnotationCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AnnotationCreate = {
    frontend_id?: (number | null);
    type: string;
    points?: string;
    color?: string;
    label_id: number;
    data_id: number;
    project_id: number;
    labelme_data?: (string | null);
    processing_stage?: string;
    workflow_execution_id?: (number | null);
    node_execution_id?: (number | null);
    annotation_id?: (number | null);
};